moved tests to tests/src folder, updated necessary files
