+++ tests/src/keybase/ecdsa.spec.ts
@@ -5,9 +5,9 @@
  *
  */
 import { expect } from "chai"
-import { Keybase } from "../../src/keybase/keybase"
-import { Account } from "../../src/models/account"
-import { InMemoryKVStore, typeGuard } from "../../src"
+import { Keybase } from "../../../src/keybase/keybase"
+import { Account } from "../../../src/models/account"
+import { InMemoryKVStore, typeGuard } from "../../../src"
 
 describe("Keybase Digital signature operations", () => {
     describe("Success scenarios", () => {
