+++ tests/src/routing.ts
@@ -5,11 +5,11 @@
 // Config
 // Constants
 import { expect } from 'chai'
-import { Pocket } from '../src/pocket'
-import { Routing } from "../src/models/routing"
-import { Node } from "../src/models/node"
-import { BondStatus } from "../src/models/output/bond-status.js"
-import { Configuration } from "../src/models/configuration.js"
+import { Pocket } from '../../src/pocket'
+import { Routing } from "../../src/models/routing"
+import { Node } from "../../src/models/node"
+import { BondStatus } from "../../src/models/output/bond-status.js"
+import { Configuration } from "../../src/models/configuration.js"
 import { PocketAAT } from "pocket-aat-js"
 
 // For Testing we are using dummy data, none of the following information is real.
